[refactor/#20] 이미지 경로 수정 파일
