+++ src/components/Button/DropDownButton.tsx
@@ -1,6 +1,6 @@
 // components/DropDownButton.tsx
 import { useState } from "react";
-import dropdownicon from "/src/assets/images/dropdown_icon.svg";
+import dropdownicon from "../../../public/icons/dropdowndown_icon.svg";
 
 interface DropDownButtonProps {
   options: string[];
