+++ src/components/TemplateWrite/CategoryTag.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import CategoryTagModal from "../../pages/TempWrite/CategoryTagModal";
 import { allTags } from "../../pages/TempWrite/tagData";
-import searchIcon2 from "../../assets/images/searchicon2.svg";
+import searchIcon2 from "../../../public/icons/searchicon2.svg";
 
 interface CategoryTagProps {
   value: string[];
