+++ src/components/TemplateWrite/EditorBlock.tsx
@@ -96,7 +96,7 @@ const EditorBlock = ({
         {block.checklistItems.length > 0 && (
           <h3 className="text-base font-semibold text-gray4 flex items-center gap-2">
             <img
-              src="/src/assets/images/alerticon.svg"
+              src="/public/icons/alerticon.svg"
               alt="alert icon"
               className="w-5 h-5"
             />
